Fikser circular reference.
